Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int s;
+    float s;
     scanf("%f",s);
     printf("%f", s);
     return 0;
